Word-valued refinement variables, APPLY lockdown
This changes it so that when a refinement is used in a function,
the value of that refinement is not #[true] but rather an unbound
WORD! which has the same symbol as the refinement itself.
If the refinement is unused, it is NONE! as before.  A legacy
mode for switching back to using TRUE is provided, but the
property will only apply to functions which were created since
the mode was turned on.

This is part (but not all) of a plan for making chaining smoother.
Another small part of this plan is included which lets you use
NONE! and GET-WORD! in refinement dispatch.  So you can
write `x: 'only` and `append/:x [a b c] [d e]` => `[a b c [d e]]`.
This is just a start, but work will be needed on path dispatch
to allow for parentheses, and just generally fixing mechanisms
so that things like `apo: :append/only` can work for capturing
refined functions into new functions.

Getting this to work required digging into APPLY, which had
two parallel codepaths with varying degrees of weakness.  It
seemed time to unify them and get the implementations to
do better checking, which allowed them to (for instance)
properly type-check the handler functions used by TRAP and
CATCH.  The mechanics were made more strict overall, so
that invoking a function through APPLY has to follow the
same basic rules as calling functions from the evaluator (e.g.
you can't set a refinement argument to true and then give
one of its arguments without the other).

Token APPLYs were switched to use the new chaining form,
but really being able to use it will require the ability for
passing NONE! to a refinement argument to "un-ask" for that
argument.